+++ crates/biome_js_analyze/tests/specs/nursery/noSkippedTests/valid.js
@@ -1,0 +1,7 @@
+describe('foo', () => {});
+it('foo', () => {});
+test('foo', () => {});
+
+describe.only('bar', () => {});
+it.only('bar', () => {});
+test.only('bar', () => {});